pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-mem0"
version = "0.1.0"
description = "Cross-session memory system for AI Agent tools, powered by mem0 + Qdrant + MCP"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "aishu" },
]
keywords = ["ai", "memory", "mcp", "mem0", "claude", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "mem0ai>=2.0.0,<3.0.0",
    "mcp>=1.0.0",
    "qdrant-client>=1.9.0",
    "ollama>=0.4.0,<1.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
litellm = ["litellm>=1.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
agent-mem0 = "agent_mem0.cli:main"

[project.urls]
Homepage = "https://github.com/aishu/agent-mem0"
Repository = "https://github.com/aishu/agent-mem0"
Issues = "https://github.com/aishu/agent-mem0/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_mem0"]

[tool.hatch.build.targets.sdist]
include = [
    "src/agent_mem0/",
    "templates/",
]
